modifications to the app

+++ ArticlesProject/php/list.php
@@ -19,17 +19,16 @@ $result = $SQL->fetchAll();
 //var_dump($_SESSION);
 
 if(isset($_SESSION["loggedin"]) && $_SESSION["loggedin"] == true)
-    echo "<nav class='navbar navbar-expand-lg navbar-light bg-light fixed-top justify-content-end'>";
-    echo "<h4>Welcome back, ".$_SESSION['username'].".</h4>";
-    if ($_SESSION['role'] === 'content_editor') {echo "<p class='flexy'><a class='btn btn-outline-success m-2' href='new.php'>Create new article</a></p>";}
-    echo "<a class='btn btn-outline-success m-2' href='search_form.php'>Search for an article</a>";
+    echo "<nav class='navbar navbar-expand-lg navbar-light bg-light fixed-top justify-content-between'>";
+    echo "<h4>Welcome back, ".$_SESSION['username'].".</h4><div class='d-flex justify-content-end'>";
+    if ($_SESSION['role'] === 'content_editor') {echo "<p class='flexy'><a class='btn btn-outline-primary m-2' href='new.php'>Create new article</a></p>";}
+    echo "<p class='flexy'><a class='btn btn-outline-primary m-2' href='search_form.php'>Search for an article</a></p>";
     echo "
-    <form method='post'>
+    <form method='post' style='height: 54px'>
     <input type='hidden' name='logout' value='123'>
-        <button type='submit' class='btn btn-outline-success m-2' name='any'>Logout</button>
-       </form>
-    <br>";
-    if ($_SERVER["REQUEST_METHOD"] == "POST" && $_POST['logout'] === '123') {
+        <button type='submit' class='btn btn-outline-primary m-2' name='any'>Logout</button>
+       </form></div>";
+    if ($_SERVER["REQUEST_METHOD"] == "POST" && (isset($_POST['logout']))) {
         session_unset();
         header("location: login.php");
     }
@@ -38,7 +37,9 @@ if(isset($_SESSION["loggedin"]) && $_SESSION["loggedin"] == true)
     echo "</nav>";
 
     //echo "<div class='row'>";
+    echo "<br><br><br>";
 
+    echo "<div class='d-flex flex-wrap'>";
 for ($count = 0; $count < count($result); $count++) {
   
 	if(is_array($result[$count]) == true ) {
@@ -47,29 +48,41 @@ for ($count = 0; $count < count($result); $count++) {
         //echo "<div class='column'>";
         //echo "<div class='hovereffect'>";
 	    if ($_SESSION['role'] === 'reader') {
-            echo "<article class='wrapper'>
-            <a href='view.php?id=".$result[$count]['id']."'>"."<h2>".$result[$count]['title']."</h2></a>".
-                "<img src='".$result[$count]['img']."'>".
-                "<p>".$result[$count]['description']."</p></article>";
+            echo "<div class='card w-25'>
+                    <div><img class='card-img-top' src='".$result[$count]['img']."'>
+                    </div>
+                    <div class='card-body'>
+                        <a href='view.php?id=".$result[$count]['id']."'>"."<h3 class='card-title'>".$result[$count]['title']."</h3></a>".
+                        "<p>".substr($result[$count]['description'], 0, 100)."...</p>
+                    </div>
+                 </div>";
         } else {
-        echo "<article class='wrapper'>
-        <a href='view.php?id=".$result[$count]['id']."'>"."<h2>".$result[$count]['title']."</h2></a>"."<img src='".$result[$count]['img']."'>".
-        "<p>".$result[$count]['description']."</p></article>
-        <form action='".$_SERVER["PHP_SELF"]."' method='post'>
-        <div class='custom-control custom-checkbox my-1 mr-sm-2' id='delete'>
-        <button type='submit' class='btn btn-primary my-1' name='delete' value='".$result[$count]['id']."'>Delete article</button>
-        </div></form>";
+        echo "<div class='card w-25'>
+                <div><img class='card-img-top' src='".$result[$count]['img']."'>
+                </div>
+                <div class='del_biathlette'>
+                    <form action='".$_SERVER["PHP_SELF"]."' method='post'>
+                        <div class='custom-control custom-checkbox my-1 mr-sm-2' id='delete'>
+                            <button type='submit' class='btn btn-primary my-1' name='delete' value='".$result[$count]['id']."'>Delete article</button>
+                        </div>
+                    </form>
+                </div>
+                <div class='card-body'>
+                    <a href='view.php?id=".$result[$count]['id']."'>"."<h3>".$result[$count]['title']."</h3></a>".
+                    "<p>".substr($result[$count]['description'], 0, 100)."...</p>
+                </div>
+             </div>";
 	}
     }
 
 }
-    //echo "</div></div>";
-if ($_SERVER["REQUEST_METHOD"] == "POST") {
+    echo "</div>";
+if ($_SERVER["REQUEST_METHOD"] == "POST" && (isset($_POST['delete']))) {
     //echo $_POST;
     $SQL = $connection->prepare('DELETE FROM article WHERE id=:Id');
     $SQL->bindParam(':Id', $_POST['delete'], PDO::PARAM_INT);
     $SQL->execute();
-    //header("location: list.php");
+    header("location: list.php");
 }
 }
 ?>
